+++ src/app/search/search-utils.test.ts
@@ -7,28 +7,35 @@ describe('generateSuggestionsForFilter', () => {
   const cases: [
     format: FilterDefinition['format'],
     keywords: FilterDefinition['keywords'],
-    suggestions: FilterDefinition['suggestions']
+    suggestions: FilterDefinition['suggestions'],
+    overload: { [key: string]: number } | undefined
   ][] = [
-    [undefined, ['a', 'b', 'c'], undefined],
-    ['query', 'a', ['b', 'c']],
-    ['stat', 'a', ['b', 'c']],
-    ['range', 'a', undefined],
-    ['rangeoverload', 'a', ['b', 'c']],
-    ['freeform', 'a', ['b', 'c']],
-    [undefined, ['a'], undefined],
-    ['stat', 'stat', allStatNames],
-    ['query', 'maxstatvalue', searchableArmorStatNames],
-    ['query', 'maxstatvalue', searchableArmorStatNames],
-    [['range', 'query'], 'energycapacity', energyCapacityTypeNames],
+    [undefined, ['a', 'b', 'c'], undefined, undefined],
+    ['query', 'a', ['b', 'c'], undefined],
+    ['stat', 'a', ['b', 'c'], undefined],
+    ['range', 'a', undefined, undefined],
+    ['range', 'a', undefined, { worthy: 10, arrivals: 11 }],
+    ['freeform', 'a', ['b', 'c'], undefined],
+    [undefined, ['a'], undefined, undefined],
+    ['stat', 'stat', allStatNames, undefined],
+    ['query', 'maxstatvalue', searchableArmorStatNames, undefined],
+    ['query', 'maxstatvalue', searchableArmorStatNames, undefined],
+    [['range', 'query'], 'energycapacity', energyCapacityTypeNames, undefined],
   ];
 
   test.each(cases)(
-    "full suggestions for filter format '%s', keyword '%s' with suggestions %s",
-    (format: FilterDefinition['format'], keywords: string | string[], suggestions?: string[]) => {
+    "full suggestions for filter format '%s', keyword '%s' with suggestions %s, overload %s",
+    (
+      format: FilterDefinition['format'],
+      keywords: string | string[],
+      suggestions?: string[],
+      overload?: { [key: string]: number } | undefined
+    ) => {
       const candidates = generateSuggestionsForFilter({
         format,
         keywords,
         suggestions,
+        overload,
       });
       expect(candidates).toMatchSnapshot();
     }