+++ config/webpack.js
@@ -426,7 +426,7 @@ module.exports = (env) => {
         // Warn when DIM Sync is off and you save some DIM-specific data
         '$featureFlags.warnNoSync': JSON.stringify(true),
         // Expose the "Add required stat mods" Loadout Optimizer toggle
-        '$featureFlags.loAutoStatMods': JSON.stringify(!env.release),
+        '$featureFlags.loAutoStatMods': JSON.stringify(true),
         // Whether to send cookies to the Bungie.net API
         '$featureFlags.apiCookies': JSON.stringify(false),
       }),