+++ src/app/loadout-builder/filter/EnergyOptions.tsx
@@ -1,60 +1,9 @@
 import { AssumeArmorMasterwork, LockArmorEnergyType } from '@destinyitemmanager/dim-api-types';
-import { PressTip } from 'app/dim-ui/PressTip';
 import { t } from 'app/i18next-t';
-import clsx from 'clsx';
-import React, { Dispatch, useMemo } from 'react';
+import { Dispatch, useMemo } from 'react';
 import { LoadoutBuilderAction } from '../loadout-builder-reducer';
 import styles from './EnergyOptions.m.scss';
-
-interface Option {
-  label: string;
-  tooltip: string;
-  selected: boolean;
-  onChange(): void;
-}
-
-const RadioSetting = React.memo(function RadioSetting({
-  label,
-  name,
-  options,
-}: {
-  label: string;
-  name: string;
-  options: Option[];
-}) {
-  return (
-    <div className={styles.settingGroup}>
-      <div className={styles.title}>{label}</div>
-      <div className={styles.buttons}>
-        {options.map(({ label, selected, tooltip, onChange }) => (
-          <RadioButton
-            key={label}
-            label={label}
-            tooltip={tooltip}
-            selected={selected}
-            onChange={onChange}
-            name={name}
-          />
-        ))}
-      </div>
-    </div>
-  );
-});
-
-function RadioButton({ label, tooltip, name, selected, onChange }: Option & { name: string }) {
-  return (
-    <PressTip
-      tooltip={tooltip}
-      elementType="label"
-      className={clsx(styles.button, {
-        [styles.selected]: selected,
-      })}
-    >
-      <input type="radio" name={name} checked={selected} onChange={onChange} />
-      {label}
-    </PressTip>
-  );
-}
+import { Option, RadioSetting } from './RadioSetting';
 
 export default function EnergyOptions({
   assumeArmorMasterwork,